requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
openai==1.3.7
python-dotenv==1.0.0
schedule==1.2.0
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
aiohttp==3.9.1
asyncio==3.4.3
python-dateutil==2.8.2
pytz==2023.3
jinja2==3.1.2
markdown==3.5.1
openpyxl==3.1.2
undetected-chromedriver==3.5.4 